Rename webpack common config
