+++ freesound/static/webpack/webpack.common.js
@@ -5,7 +5,7 @@ const SUPPORTED_BROWSERS = ['>1%', 'last 4 versions', 'Firefox ESR', 'not ie < 1
 module.exports = {
   entries: {
     base: path.join(__dirname, '..', 'base.js'),
-    index: path.join(__dirname, '..', 'src/index.js'),
+    index: path.join(__dirname, '..', 'src/common.js'),
   },
   output: {
     filename: '[name].js',
@@ -51,5 +51,15 @@ module.exports = {
         ],
       },
     },
+    fileLoader: {
+      test: /\.svg$/,
+      exclude: /icons/,
+      loaders: ['file-loader']
+    },
+    iconsLoader: {
+      test: /\.svg$/,
+      include: /icons/,
+      loaders: ['raw-loader']
+    }
   },
 };
